test(geth): isolate opcode traversal regression coverage
